requirements.txt
discord
asyncio
python-dotenv
pandas